Change syntax for mass matrix derivatives
To support anisotropic PDE coefficients the mass matrix functions
can be implemented for coefficient vectors of variable lengths,
providing, e.g., the x,y,z components of a diagonal tensor.
Consequently, the derivatives must be able to distinguish these
cases. This is achieved by adding the PDE coefficient as an input
argument to the mass matrix derivatives.